requirements.txt
Flask
google-generativeai
Flask-SQLAlchemy
psycopg2-binary